Move AtomicOrbitals back into core via a SpheriCart extension
Replaces the lib/AtomicOrbitals subpackage: the type and all Ylm-generic
logic live in src/atomicorbitals/, with the SpheriCart-specific glue
(complex value type, _rand_* fixtures) in ext/SpheriCartExt.jl behind a
[weakdeps] SpheriCart. Avoids registering a generically-named package.
