+++ Project.toml
@@ -25,6 +25,12 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 WithAlloc = "fb1aa66a-603c-4c1d-9bc4-66947c7b08dd"
 
+[weakdeps]
+SpheriCart = "5caf2b29-02d9-47a3-9434-5931c85ba645"
+
+[extensions]
+SpheriCartExt = "SpheriCart"
+
 [compat]
 ACEbase = "0.4.7"
 BenchmarkTools = "1"
@@ -43,6 +49,7 @@ QuadGK = "2"
 Random = "1.10"
 SparseArrays = "1.10"
 SpecialFunctions = "2.2"
+SpheriCart = "0.2.4"
 StaticArrays = "1.5"
 Test = "1.10"
 WithAlloc = "0.1.0"
@@ -51,8 +58,9 @@ julia = "1.10"
 [extras]
 Optimisers = "3bd65402-5787-11e9-1adc-39752487f4e2"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
+SpheriCart = "5caf2b29-02d9-47a3-9434-5931c85ba645"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [targets]
-test = ["Test", "Printf", "Optimisers", "Zygote"]
+test = ["Test", "Printf", "Optimisers", "Zygote", "SpheriCart"]
