+++ src/atomicorbitals/radialdecay.jl
@@ -15,7 +15,7 @@ df(f::SlaterDecay, x::T) where T = one(T)
 """
     construct_basis(RadialDecay, ζ_raw, D_raw, decay)
 
-Construct a `RadialDecay` object from raw matrix data `ζ_raw`, `D_raw` 
+Construct a `RadialDecay` object from raw matrix data `ζ_raw`, `D_raw`
 and a `DecayFunction(f, df)` representing the decay form and its derivative.
 All input is converted to statically-sized `SMatrix` for efficiency.
 """
@@ -32,21 +32,21 @@ Base.length(basis::RadialDecay) = size(basis.ζ, 1)
 
 _valtype(::RadialDecay, T::Type{<: Number}, args...) = T
 
-_valtype(::RadialDecay, T::Type{<: Real}, 
+_valtype(::RadialDecay, T::Type{<: Real},
          ps, st) = promote_type(T, eltype(ps.ζ), eltype(ps.D))
 
 
 _static_params(basis::RadialDecay) = (ζ = basis.ζ, D = basis.D)
 
-_init_luxparams(basis::RadialDecay) = 
+_init_luxparams(basis::RadialDecay) =
             ( ζ = Matrix(basis.ζ), D = Matrix(basis.D) )
 
-_evaluate!(P, dP, basis::RadialDecay, x)  = 
+_evaluate!(P, dP, basis::RadialDecay, x)  =
     _evaluate!(P, dP, basis, x, _static_params(basis), nothing)
 
 natural_indices(basis::RadialDecay) = basis.spec
 
-function _evaluate!(P, dP, basis::RadialDecay, x::AbstractVector, ps, st) 
+function _evaluate!(P, dP, basis::RadialDecay, x::AbstractVector, ps, st)
     ζ, D = ps.ζ, ps.D
     N, K = size(ζ)
     nX = length(x)
@@ -59,12 +59,12 @@ function _evaluate!(P, dP, basis::RadialDecay, x::AbstractVector, ps, st)
 
     decay = basis.decay
 
-    @inbounds begin 
+    @inbounds begin
         for n = 1:N, m = 1:K
-            @simd ivdep for i = 1:nX 
+            @simd ivdep for i = 1:nX
                 fx = decay(x[i])
                 a = D[n, m] * exp(-ζ[n, m] * fx)
-                P[i, n] += a 
+                P[i, n] += a
                 if WITHGRAD
                     dfx = df(decay, x[i])
                     dP[i, n] += -ζ[n, m] * dfx * a
@@ -73,7 +73,7 @@ function _evaluate!(P, dP, basis::RadialDecay, x::AbstractVector, ps, st)
         end
     end
 
-    return nothing 
+    return nothing
 end
 
 function pullback_ps(∂P, basis::RadialDecay, x::BATCH, ps, st)
@@ -86,7 +86,7 @@ function pullback_ps(∂P, basis::RadialDecay, x::BATCH, ps, st)
     ∂D = fill!(similar(D), 0)
 
     @inbounds for n = 1:N, m = 1:K
-        @simd ivdep for j = 1:nX 
+        @simd ivdep for j = 1:nX
             fx  = decay(x[j])           # f(x[j])
             dfx = df(decay, x[j])       # df/dx
             a1  = exp(-ζ[n, m] * fx)
@@ -97,38 +97,19 @@ function pullback_ps(∂P, basis::RadialDecay, x::BATCH, ps, st)
     return (ζ = ∂ζ, D = ∂D)
 end
 
-function _rand_basis(N1=4, N2=3; 
-    K::Int=1, 
-    T::Type=Float64, 
-    decay_type::AbstractDecayFunction=GaussianDecay(),
-    ζinit = () -> rand(T, N1 * N2 * N1^2, K), 
-    Dinit = () -> ones(T, N1 * N2 * N1^2, K))
-
-    Pn = MonoBasis(N1 + 1)
-    Ylm = SolidHarmonics(N1 - 1)
-    spec_list = [(n1=n1, n2=n2, l=l, m=m) for n1 in 1:N1, n2 in 1:N2, l in 0:N1-1 for m in -l:l]
-    spec = SVector{length(spec_list)}(spec_list)
-    spec_ln = unique((n1=s.n1, n2=s.n2, l=s.l) for s in spec)
-    Dn = construct_basis(ζinit(), Dinit(), decay_type, spec_ln)
-    specidx = _specidx(spec, Pn, Dn, Ylm)
-
-    return AtomicOrbitals{length(spec), typeof(Pn), typeof(Dn), typeof(Ylm)}(Pn, Dn, Ylm, spec, specidx)
-end
-
-_rand_gaussian_basis(N1=4, N2=3, T=Float64) = _rand_basis(N1, N2; T=T)
-
-_rand_slater_basis(N1=4, N2=3, T=Float64) = _rand_basis(N1, N2; T=T, decay_type = SlaterDecay())
-
-_rand_sto_basis(N1=4, N2=2, K=4, T=Float64) = _rand_basis(N1, N2; T=T, K=K, 
-        ζinit = () -> rand(T, N1 * N2 * N1^2, K),
-        Dinit = () -> rand(T, N1 * N2 * N1^2, K))
+# `_rand_*` are test/dev fixtures that default the angular part to a SpheriCart
+# `SolidHarmonics`; their methods are defined in ext/SpheriCartExt.jl.
+function _rand_basis end
+function _rand_gaussian_basis end
+function _rand_slater_basis end
+function _rand_sto_basis end
 
 function _invmap(a::AbstractVector)
     inva = Dict{eltype(a), Int}()
-    for i = 1:length(a) 
-       inva[a[i]] = i 
+    for i = 1:length(a)
+       inva[a[i]] = i
     end
-    return inva 
+    return inva
 end
 
 function _specidx(spec, Pn, Dn, Ylm)
@@ -139,6 +120,6 @@ function _specidx(spec, Pn, Dn, Ylm)
     spec_Dn = natural_indices(Dn); inv_Dn = _invmap(spec_Dn)
     for (z, b) in enumerate(spec)
         specidx[z] = (inv_Pn[(n = b.n1, )], inv_Dn[(n1 = b.n1, n2 = b.n2, l = b.l)], inv_Ylm[(l=b.l, m=b.m)])
-    end  
+    end
     return specidx
 end
