+++ test/test_atorbrad.jl
@@ -3,13 +3,12 @@ using StaticArrays
 using LinearAlgebra
 using LuxCore
 import Polynomials4ML as P4ML
-import AtomicOrbitals
-import SpheriCart, ACEbase   # ensure the SpheriCart ACEbase extension is active
+import SpheriCart, ACEbase   # activate the SpheriCart ACEbase + P4ML SpheriCart extensions
 using Polynomials4ML: evaluate, evaluate_ed
 
 
 @testset "GaussianBasis + HyperDual matches evaluate/evaluate_ed" begin
-    basis = AtomicOrbitals._rand_gaussian_basis()
+    basis = P4ML._rand_gaussian_basis()
     P4ML.Testing.test_hyperdual_consistency(basis)
     P4ML.Testing.test_evaluate_xx(basis)
     P4ML.Testing.test_chainrules(basis)
@@ -17,7 +16,7 @@ using Polynomials4ML: evaluate, evaluate_ed
 end
 
 @testset "SlaterBasis + HyperDual matches evaluate/evaluate_ed" begin
-    basis = AtomicOrbitals._rand_slater_basis()
+    basis = P4ML._rand_slater_basis()
     P4ML.Testing.test_hyperdual_consistency(basis)
     P4ML.Testing.test_evaluate_xx(basis)
     P4ML.Testing.test_chainrules(basis)
@@ -25,7 +24,7 @@ end
 end
 
 @testset "STOBasis + HyperDual matches evaluate/evaluate_ed" begin
-    basis = AtomicOrbitals._rand_sto_basis()
+    basis = P4ML._rand_sto_basis()
     P4ML.Testing.test_hyperdual_consistency(basis)
     P4ML.Testing.test_evaluate_xx(basis)
     P4ML.Testing.test_chainrules(basis)
